int to ulong error convertion bug fixed

+++ Api/Services/Transaction/ITransactionService.cs
@@ -5,8 +5,8 @@ namespace Api.Services.Transaction
 {
     public interface ITransactionService
     {
-        Task<CardanoSharp.Wallet.Models.Transactions.Transaction> BuildTransaction(string fromAddress, string toAddress, int value);
+        Task<CardanoSharp.Wallet.Models.Transactions.Transaction> BuildTransaction(string fromAddress, string toAddress, ulong value);
         Task<CardanoSharp.Wallet.Models.Transactions.Transaction> SignTransaction(string transactionCbor, string witness);
-        Task<ulong> CalculateFee(string fromAddress, string toAddress, int value);
+        Task<ulong> CalculateFee(string fromAddress, string toAddress, ulong value);
     }
 }
