+++ Api/Transaction/TxBuilder.cs
@@ -36,7 +36,7 @@ namespace Api.Transaction
 
             try
             {
-                int transferValue = int.Parse(value);
+                ulong transferValue = ulong.Parse(value);
                 var transact = await _transaction.BuildTransaction(walletFrom,walletTo, transferValue);
                 return new OkObjectResult(transact);
             }
